Rename hook and hook directory
Renames the `customhooks` directory to `hooks`
and the `RoleLoggedIn` hook to `useRoleLoggedIn`.
